move mol mass test to new test module
