+++ molecool/tests/test_molecool.py
@@ -50,13 +50,3 @@ def test_build_bond_list():
     bonds = molecool.build_bond_list(coordinates)
 
     assert len(bonds) == 4
-
-def test_molecular_mass():
-    symbols = ['C', 'H', 'H', 'H', 'H']
-
-    calculated_mass = molecool.calculate_molecular_mass(symbols)
-
-    actual_mass = 16.04
-
-    assert pytest.approx(actual_mass, abs=1e-2) == calculated_mass
-
