+++ molecool/tests/test_molecule.py
@@ -1,0 +1,16 @@
+"""
+Tests for the molecule module.
+"""
+
+import pytest
+import molecool
+
+def test_molecular_mass():
+    symbols = ['C', 'H', 'H', 'H', 'H']
+
+    calculated_mass = molecool.calculate_molecular_mass(symbols)
+
+    actual_mass = 16.04
+
+    assert pytest.approx(actual_mass, abs=1e-2) == calculated_mass
+
